pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.20.0.post0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.9.1",
    "pytest-playwright>=0.7.0",
    "ruff>=0.11.13",
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
